[ContentListing] - Refactor: move some files to a new, dedicated package
