+++ content_listing/src/main/kotlin/edu/artic/decoration/AllExhibitionsItemDecoration.kt
@@ -1,4 +1,4 @@
-package edu.artic.exhibitions.recyclerview
+package edu.artic.decoration
 
 import android.content.Context
 import android.graphics.Rect
