+++ content_listing/src/main/kotlin/edu/artic/decoration/AllToursItemDecoration.kt
@@ -1,4 +1,4 @@
-package edu.artic.tours.recyclerview
+package edu.artic.decoration
 
 import android.content.Context
 import android.graphics.Rect
