+++ content_listing/src/main/kotlin/edu/artic/exhibitions/AllExhibitionsFragment.kt
@@ -13,7 +13,7 @@ import edu.artic.adapter.itemClicksWithPosition
 import edu.artic.analytics.ScreenName
 import edu.artic.base.utils.asDeepLinkIntent
 import edu.artic.content.listing.R
-import edu.artic.exhibitions.recyclerview.AllExhibitionsItemDecoration
+import edu.artic.decoration.AllExhibitionsItemDecoration
 import edu.artic.navigation.NavigationConstants
 import edu.artic.viewmodel.BaseViewModelFragment
 import edu.artic.viewmodel.Navigate
